+++ modules/apps/headless/headless-collaboration/headless-collaboration-api/src/main/java/com/liferay/headless/collaboration/resource/v1_0/BlogPostingImageResource.java
@@ -41,10 +41,6 @@ public interface BlogPostingImageResource {
 	public BlogPostingImage getBlogPostingImage(Long blogPostingImageId)
 		throws Exception;
 
-	public BlogPostingImage patchBlogPostingImage(
-			Long blogPostingImageId, MultipartBody multipartBody)
-		throws Exception;
-
 	public BlogPostingImage putBlogPostingImage(
 			Long blogPostingImageId, MultipartBody multipartBody)
 		throws Exception;
