+++ modules/apps/headless/headless-collaboration/headless-collaboration-impl/src/main/java/com/liferay/headless/collaboration/internal/resource/v1_0/BaseBlogPostingImageResourceImpl.java
@@ -45,7 +45,6 @@ import javax.validation.constraints.NotNull;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
-import javax.ws.rs.PATCH;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
@@ -85,21 +84,6 @@ public abstract class BaseBlogPostingImageResourceImpl
 	public BlogPostingImage getBlogPostingImage(
 			@NotNull @PathParam("blog-posting-image-id") Long
 				blogPostingImageId)
-		throws Exception {
-
-		return new BlogPostingImage();
-	}
-
-	@Override
-	@Consumes("multipart/form-data")
-	@PATCH
-	@Path("/blog-posting-images/{blog-posting-image-id}")
-	@Produces("application/json")
-	@Tags(value = {@Tag(name = "BlogPostingImage")})
-	public BlogPostingImage patchBlogPostingImage(
-			@NotNull @PathParam("blog-posting-image-id") Long
-				blogPostingImageId,
-			MultipartBody multipartBody)
 		throws Exception {
 
 		return new BlogPostingImage();
