+++ modules/apps/headless/headless-collaboration/headless-collaboration-test/src/testIntegration/java/com/liferay/headless/collaboration/resource/v1_0/test/BaseBlogPostingImageResourceTestCase.java
@@ -242,69 +242,6 @@ public abstract class BaseBlogPostingImageResourceTestCase {
 	}
 
 	@Test
-	public void testPatchBlogPostingImage() throws Exception {
-		Assert.assertTrue(true);
-	}
-
-	protected BlogPostingImage testPatchBlogPostingImage_addBlogPostingImage(
-			BlogPostingImage blogPostingImage)
-		throws Exception {
-
-		throw new UnsupportedOperationException(
-			"This method needs to be implemented");
-	}
-
-	protected BlogPostingImage invokePatchBlogPostingImage(
-			Long blogPostingImageId, MultipartBody multipartBody)
-		throws Exception {
-
-		Http.Options options = _createHttpOptions();
-
-		String location =
-			_resourceURL +
-				_toPath(
-					"/blog-posting-images/{blog-posting-image-id}",
-					blogPostingImageId);
-
-		options.setLocation(location);
-
-		options.setPatch(true);
-
-		String string = HttpUtil.URLtoString(options);
-
-		try {
-			return _outputObjectMapper.readValue(
-				string, BlogPostingImage.class);
-		}
-		catch (Exception e) {
-			Assert.fail("HTTP response: " + string);
-
-			throw e;
-		}
-	}
-
-	protected Http.Response invokePatchBlogPostingImageResponse(
-			Long blogPostingImageId, MultipartBody multipartBody)
-		throws Exception {
-
-		Http.Options options = _createHttpOptions();
-
-		String location =
-			_resourceURL +
-				_toPath(
-					"/blog-posting-images/{blog-posting-image-id}",
-					blogPostingImageId);
-
-		options.setLocation(location);
-
-		options.setPatch(true);
-
-		HttpUtil.URLtoString(options);
-
-		return options.getResponse();
-	}
-
-	@Test
 	public void testPutBlogPostingImage() throws Exception {
 		Assert.assertTrue(true);
 	}
