+++ modules/apps/headless/headless-collaboration/headless-collaboration-test/src/testIntegration/java/com/liferay/headless/collaboration/resource/v1_0/test/BaseBlogPostingResourceTestCase.java
@@ -953,14 +953,6 @@ public abstract class BaseBlogPostingResourceTestCase {
 			return sb.toString();
 		}
 
-		if (entityFieldName.equals("caption")) {
-			sb.append("'");
-			sb.append(String.valueOf(blogPosting.getCaption()));
-			sb.append("'");
-
-			return sb.toString();
-		}
-
 		if (entityFieldName.equals("contentSpaceId")) {
 			throw new IllegalArgumentException(
 				"Invalid entity field " + entityFieldName);
@@ -1036,11 +1028,6 @@ public abstract class BaseBlogPostingResourceTestCase {
 				"Invalid entity field " + entityFieldName);
 		}
 
-		if (entityFieldName.equals("imageId")) {
-			throw new IllegalArgumentException(
-				"Invalid entity field " + entityFieldName);
-		}
-
 		if (entityFieldName.equals("keywords")) {
 			throw new IllegalArgumentException(
 				"Invalid entity field " + entityFieldName);
@@ -1070,7 +1057,6 @@ public abstract class BaseBlogPostingResourceTestCase {
 			{
 				alternativeHeadline = RandomTestUtil.randomString();
 				articleBody = RandomTestUtil.randomString();
-				caption = RandomTestUtil.randomString();
 				contentSpaceId = RandomTestUtil.randomLong();
 				dateCreated = RandomTestUtil.nextDate();
 				dateModified = RandomTestUtil.nextDate();
@@ -1081,7 +1067,6 @@ public abstract class BaseBlogPostingResourceTestCase {
 				hasComments = RandomTestUtil.randomBoolean();
 				headline = RandomTestUtil.randomString();
 				id = RandomTestUtil.randomLong();
-				imageId = RandomTestUtil.randomLong();
 			}
 		};
 	}
